// frontend-fap/src/app/tags/models/create.tag.ts
/**
 * FilmArchivePlanner (FAP)
 * REST API for the FAP backend.
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CreateUpdateTagDto {
  value: string;
  type: string;
  user: string;
  public: boolean;
}